requirements/tests.txt
flake8==2.2.2
django-discover-runner==1.0
coverage==3.7.1
coveralls==0.4.2
